--- a/onnxruntime/test/mlas/bench/bench_util.h
+++ b/onnxruntime/test/mlas/bench/bench_util.h
@@ -8,12 +8,6 @@
 #include <functional>
 #include <random>
 
-// Specifies benchmark arguments from the cartesian product of `arglists`, like Benchmark::ArgsProduct().
-// `include_filter` is called to determine whether a given set of arguments should be included.
-void ArgsProductWithFilter(benchmark::internal::Benchmark* bench,
-                           const std::vector<std::vector<int64_t>>& arglists,
-                           std::function<bool(const std::vector<int64_t>& args)> include_filter);
-
 template <typename ElementType>
 std::vector<ElementType> RandomVectorUniform(
     size_t N,