### Description
<!-- Describe your changes. -->

Update SQ4BitGemm CompInt8 implementation to process multiple blocks along a single column instead of processing single blocks from multiple columns.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

Performance optimization.

#### Benchmarks

M = 1, single threaded, compute type = CompInt8, symmetric quantization.

Baseline: f06164ef8b
Updated: 3b474b8342

BlkLen | K | N | baseline (ns) | updated (ns) | % improvement [(baseline - updated) / baseline * 100]
-- | -- | -- | -- | -- | --
16 | 4096 | 4096 | 1362806 | 1204807 | 11.59365
  | 11008 | 4096 | 3614324 | 3244342 | 10.23655
  | 4096 | 11008 | 3726424 | 3227573 | 13.38686
32 | 4096 | 4096 | 777730 | 646762 | 16.83978
  | 11008 | 4096 | 2186339 | 1747091 | 20.09057
  | 4096 | 11008 | 2183616 | 1753398 | 19.70209
128 | 4096 | 4096 | 670500 | 593283 | 11.51633
  | 11008 | 4096 | 1859688 | 1614690 | 13.17415
  | 4096 | 11008 | 1886570 | 1611324 | 14.58976