@yufenglee 2024-03-13
Did you try same optimization on the CompFloat32?