@edgchen1 2024-03-14
> Did you try same optimization on the CompFloat32?

Not yet. Would probably do that in another PR.